UNREAL -  this commit was a one shot prompt!!
"add what looks like a class Windows 98 "help doc" on the desktop... make it clickable so that it opens up in what looks ike classic Windows 98 "help" documentation.  add functionalty for our client to render markdown, and add a help.md file in the `/public` folder where we can manage the user documentation for this app. when user clicks the help icon it will pull the markdown doc from the /public folder and render it in a Windows 98 way"